Migrate to 64 bit, Higher half kernel with Limine Bootloader (#5)
* Basic setup done, global constructors working

* Use limine terminal for printf

* stack smashing working now

* Obtain mem map from limine and setup GDT

* ISR works now

* PMM migrated

* Update interrupt handling asm (#4)

* WIP for VMM

* VMM works now, better than ever

* new/delete and malloc/free works now

* Add limine to repo

* Update license

+++ kernel/include/kernel/memory/heap.h
@@ -5,25 +5,25 @@
 namespace Heap {
 
     struct __attribute__((packed)) Node {
-        uint32_t size;
+        uint64_t size;
         Node* next = nullptr;
         void print();
     };
 
-    constexpr int sizeofNode = sizeof(Node);
+    constexpr int64_t sizeofNode = sizeof(Node);
 
     class FreeList {
         Node* head;
     public:
         FreeList() { head = nullptr; }
-        FreeList(VirtualAddr addr, uint32_t size);
-        void* alloc(uint32_t size);
-        void addMemory(VirtualAddr addr, uint32_t size);
+        FreeList(VirtualAddr addr, uint64_t size);
+        void* alloc(uint64_t size);
+        void addMemory(VirtualAddr addr, uint64_t size);
         void free(void* ptr);
         void print();
     private:
         // returns the other part of memory
-        Node* split(Node* node, uint32_t size);
+        Node* split(Node* node, uint64_t size);
         void addNode(Node* newNode);
     };
 };
